DRY up hook, add to profile updates too
